Beginner Use case
Working example (with instructions) to create a Pluralsight Lab with a AWS sandbox region. The sandbox comes with Terraform pre baked on a EC2 instance.
Using this Terraform installed, create a main.tf file. This will allow to connect to required AZ subnet and launch a t2.micro EC2 instance.